fix(tasks): modify task interface to mount on dynamic mounter didn't test all of them but typescript compiles and hardhat mounts successfully
