+++ contracts/dollar/tasks/library/manager.ts
@@ -1,11 +1,13 @@
 import "@nomiclabs/hardhat-waffle";
-import { task } from "hardhat/config";
-import { UbiquityAlgorithmicDollarManager } from "../artifacts/types/UbiquityAlgorithmicDollarManager";
+import { ActionType, HardhatRuntimeEnvironment } from "hardhat/types";
+import { UbiquityAlgorithmicDollarManager } from "../../artifacts/types/UbiquityAlgorithmicDollarManager";
 // This file is only here to make interacting with the Dapp easier,
 // feel free to ignore it if you don't need it.
 
-task("manager", "Get info about manager contract's address").setAction(
-  async (taskArgs: { receiver: string; manager: string }, { ethers }) => {
+export const description = "Get info about manager contract's address";
+export const action =
+  (): ActionType<any> =>
+  async (_taskArgs, { ethers }: HardhatRuntimeEnvironment) => {
     const net = await ethers.provider.getNetwork();
     const managerAdr = "0x4DA97a8b831C345dBe6d16FF7432DF2b7b776d98";
     const debtCouponMgrAdr = "0x432120Ad63779897A424f7905BA000dF38A74554";
@@ -13,20 +15,15 @@ task("manager", "Get info about manager contract's address").setAction(
       console.warn("You are running the   task with Hardhat network");
     }
     console.log(`net chainId: ${net.chainId}  `);
-    const manager = (await ethers.getContractAt(
-      "UbiquityAlgorithmicDollarManager",
-      managerAdr
-    )) as UbiquityAlgorithmicDollarManager;
+    const manager = (await ethers.getContractAt("UbiquityAlgorithmicDollarManager", managerAdr)) as UbiquityAlgorithmicDollarManager;
     const mgrtwapOracleAddress = await manager.twapOracleAddress();
     const mgrdebtCouponAddress = await manager.debtCouponAddress();
     const mgrDollarTokenAddress = await manager.dollarTokenAddress();
     const mgrcouponCalculatorAddress = await manager.couponCalculatorAddress();
-    const mgrdollarMintingCalculatorAddress =
-      await manager.dollarMintingCalculatorAddress();
+    const mgrdollarMintingCalculatorAddress = await manager.dollarMintingCalculatorAddress();
     const mgrbondingShareAddress = await manager.bondingShareAddress();
     const mgrbondingContractAddress = await manager.bondingContractAddress();
-    const mgrstableSwapMetaPoolAddress =
-      await manager.stableSwapMetaPoolAddress();
+    const mgrstableSwapMetaPoolAddress = await manager.stableSwapMetaPoolAddress();
     const mgrcurve3PoolTokenAddress = await manager.curve3PoolTokenAddress(); // 3CRV
     const mgrtreasuryAddress = await manager.treasuryAddress();
     const mgruGOVTokenAddress = await manager.governanceTokenAddress();
@@ -36,8 +33,7 @@ task("manager", "Get info about manager contract's address").setAction(
     const mgrautoRedeemTokenAddress = await manager.autoRedeemTokenAddress(); // uAR
     const mgruarCalculatorAddress = await manager.uarCalculatorAddress(); // uAR calculator
 
-    const mgrExcessDollarsDistributor =
-      await manager.getExcessDollarsDistributor(debtCouponMgrAdr);
+    const mgrExcessDollarsDistributor = await manager.getExcessDollarsDistributor(debtCouponMgrAdr);
     console.log(`
       ****
       debtCouponMgr:${debtCouponMgrAdr}
@@ -60,5 +56,4 @@ task("manager", "Get info about manager contract's address").setAction(
       uarCalculatorAddress:${mgruarCalculatorAddress}
       ExcessDollarsDistributor:${mgrExcessDollarsDistributor}
       `);
-  }
-);
+  };
