+++ contracts/dollar/tasks/library/price-reset-i/getAddresses.ts
